Add MediaKeyService that has playlist
