+++ naver-music-for-mac/AppDelegate.swift
@@ -10,14 +10,10 @@ import Cocoa
 import RealmSwift
 import Fabric
 import Crashlytics
-import MediaKeyTap
 
 @NSApplicationMain
 class AppDelegate: NSObject, NSApplicationDelegate {
   var window: WindowController?
-  
-  // MediaKeyTap
-  var mediaKeyTab: MediaKeyTap?
   
   override func awakeFromNib() {
     super.awakeFromNib()
@@ -25,38 +21,15 @@ class AppDelegate: NSObject, NSApplicationDelegate {
                                                                     migrationBlock: { (migration, oldSchemaVersion) in
     })
     UserDefaults.standard.register(defaults: ["NSApplicationCrashOnExceptions": true])
-    Fabric.with([Crashlytics.self, Answers.self])
+//    Fabric.with([Crashlytics.self, Answers.self])
     self.window = WindowController()
     window?.showWindow(nil)
     
-    // MediaKeyTap
-    self.mediaKeyTab = MediaKeyTap(delegate: self)
-    self.mediaKeyTab?.start()
+    // MediaKey
+    MediaKeyService.shared.start()
     
   }
 
 }
 
-extension AppDelegate: MediaKeyTapDelegate {
-  func handle(mediaKey: MediaKey, event: KeyEvent) {
-    
-    switch mediaKey {
-    case .playPause:
-      let playlist = Playlist.getMyPlayList()
-      playlist.setIsPaused(isPaused: !playlist.isPaused)
-      break
-      
-    case .rewind:
-      Playlist.getMyPlayList().prev()
-      break
-      
-    case .fastForward:
-      Playlist.getMyPlayList().next()
-      break
-      
-    default:
-      break
-    }
-  }
-}
